requirements.txt
-i https://pypi.org/simple/
bcrypt==3.2.0
cffi==1.14.2
cryptography==3.1
loguru==0.5.1
paramiko==2.7.2
pycparser==2.20
pynacl==1.4.0
python-dotenv==0.14.0
scp==0.13.2
six==1.15.0
psutil==5.7.3
tabulate==0.8.7